requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Login==0.6.3
python-dotenv==1.0.0
proxmoxer==2.2.0
requests==2.31.0
paramiko==3.4.0
python-ldap==3.4.3
psycopg2-binary==2.9.9
marshmallow==3.20.1
Flask-Marshmallow==1.2.0
celery==5.3.4
redis==5.0.1
flasgger==0.9.7.1
pytest==7.4.0
pytest-mock==3.10.0
pytest-cov==4.1.0
flask-jwt-extended==4.4.4
flask-bcrypt==1.0.1